requirements.txt
packaging==24.1
pillow==10.3.0
pytesseract==0.3.10
